fix #134 - potential wrong current fleet construction
run_current in automatic solution can potentially build the wrong fleet for comparison. This will construct a vehicle usage and driving book with wrong vehicle makes and models.

Without changing the count on automatic sim, the bug could lead to wrong counts on current fleets.

Issue happens because the current fleet is build from a concatenated string of unique vehicles from the db that share attributes. To find exactly the current vehicle and populate a simulation fleet, the vehicle_id is matched with the string. The string matching is not accurate because it can match 26 on 264 - pulling out the wrong vehicle.

Splitting the unique vehicle string will make the match accurate and ensure a proper match.
fleetmanager/model/genetic.py - run_current